[autogen][py] Provide convenience functions for motor commands
Each motor class (tacho-motor, dc-motor, servo-motor) provides a set of
methods corresponding to possible commands. The methods take variable
number of keyworded parameters. Each parameter should correspond to
writable attribute of the motor. The attributes will be set before
setting the command.

The methods are generated automatically based on the list of possible
commands in spec.json.

Example:

    >>> m = large_motor()
    >>> m.run_timed(time_sp=1000, duty_cycle_sp=75)

The second line is equivalent to

    >>> m.time_sp = 1000
    >>> m.duty_cycle_sp = 75
    >>> m.command = 'run-timed'