Since the 2015-12-30 image address is used in stead of port_name (#144)
Since I'am a complete git-nitwit, i choose to re-do this patch. Will close the other pull request on
https://github.com/BertLindeman/ev3dev-lang/commit/c9ab02bc8027c0957939b7a61b9577faace7cf61

Since the 2015-12-30 image the lego/port<N> is no longer incremented. 
It might be done in the future, so I left the increment in this doc.
I am not sure if this spec is auto-generated and if so I do not know where it came from .. .. ..